fix: use fresh cart data, add thumbnail guard, use Link, and restore actions symmetry
- booking-card: capture addToCart return value for accurate cart count
- related-class-card: guard thumbnail.url access, replace <a> with <Link>
- cart actions: remove orchestration function, move logic to component
  using existing 1:1 wrappers from classes/actions
